+++ assets/javascripts/kitten/components/cards/crowdfunding-card/stories.js
@@ -77,6 +77,7 @@ export const CrowdfundingCard = () => (
       info3={text('Info3', 'Lorem ipsum…')}
       progress={number('Progress', 42)}
       state={text('State', 'Lorem ipsum…')}
+      buttonText={text('Button text', null)}
     />
   </StoryContainer>
 )